fix: size_t and uint64_t discrepancies between interfaces and implementation

+++ endpoint/store/include/privmx/endpoint/store/FileHandler.hpp
@@ -102,7 +102,7 @@ public:
     uint64_t size() override;
     void seekg(const uint64_t pos) override;
     void seekp(const uint64_t pos) override;
-    core::Buffer read(const size_t length) override;
+    core::Buffer read(const uint64_t length) override;
     void write(const core::Buffer& chunk, bool truncate = false) override;
 
     inline void close() override {}
